requirements.txt
appnope==0.1.0
bleach==2.0.0
certifi==2017.7.27.1
-e git+http://github.com/tensorflow/cleverhans.git@2950bf19a184458dbd31362e97f3e9c1e95c694f#egg=cleverhans
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
greenlet==0.4.12
h5py==2.7.0
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==2.0.8
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.1
msgpack-python==0.4.8
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.0
nose==1.3.7
notebook==5.2.1
numpy==1.12.1
olefile==0.44
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.2.0
ptyprocess==0.5.2
pycodestyle==2.3.1
pydot==1.2.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
statsmodels==0.8.0
tensorflow==1.1.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
yapf==0.20.0
